# Back-end/PVWApi/PVWApi/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "PVWApi, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Convidado": {
      "get": {
        "tags": [
          "Convidado"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Convidado"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Convidado"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Convidado"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Convidado"
            }
          }
        }
      }
    },
    "/api/Convidado/{id}": {
      "get": {
        "tags": [
          "Convidado"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Convidado"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Convidado"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Convidado"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Convidado"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": [
          "Convidado"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Operation"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Banda": {
      "type": "object",
      "properties": {
        "bandaId": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "numeroMusicas": {
          "format": "int32",
          "type": "integer"
        },
        "numeroIntegrantes": {
          "format": "int32",
          "type": "integer"
        },
        "horasContratadas": {
          "format": "date-span",
          "type": "string"
        },
        "valor": {
          "format": "double",
          "type": "number"
        },
        "festa": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Festa"
          }
        }
      }
    },
    "Buffet": {
      "type": "object",
      "properties": {
        "buffetId": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "horasContratadas": {
          "format": "date-span",
          "type": "string"
        },
        "valor": {
          "format": "double",
          "type": "number"
        },
        "comidas": {
          "type": "string"
        },
        "festa": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Festa"
          }
        }
      }
    },
    "Casamento": {
      "type": "object",
      "properties": {
        "casamentoId": {
          "format": "int32",
          "type": "integer"
        },
        "nomeParceiroA": {
          "type": "string"
        },
        "nomeParceiroB": {
          "type": "string"
        },
        "status": {
          "type": "boolean"
        },
        "data": {
          "format": "date-time",
          "type": "string"
        },
        "usuarioId": {
          "format": "int32",
          "type": "integer"
        },
        "festaId": {
          "format": "int32",
          "type": "integer"
        },
        "decoracaoId": {
          "format": "int32",
          "type": "integer"
        },
        "usuario": {
          "$ref": "#/definitions/Usuario"
        },
        "festa": {
          "$ref": "#/definitions/Festa"
        },
        "cerimonia": {
          "$ref": "#/definitions/Cerimonia"
        },
        "decoracao": {
          "$ref": "#/definitions/Decoracao"
        },
        "convidado": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Convidado"
          }
        }
      }
    },
    "Celebrante": {
      "type": "object",
      "properties": {
        "celebranteId": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "horasContratadas": {
          "format": "date-span",
          "type": "string"
        },
        "valor": {
          "format": "double",
          "type": "number"
        },
        "cerimonia": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Cerimonia"
          }
        }
      }
    },
    "Cerimonia": {
      "type": "object",
      "properties": {
        "cerimoniaId": {
          "format": "int32",
          "type": "integer"
        },
        "civil": {
          "type": "boolean"
        },
        "religiosa": {
          "type": "boolean"
        },
        "local": {
          "type": "string"
        },
        "date": {
          "format": "date-time",
          "type": "string"
        },
        "celebranteId": {
          "format": "int32",
          "type": "integer"
        },
        "celebrante": {
          "$ref": "#/definitions/Celebrante"
        },
        "casamento": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Casamento"
          }
        }
      }
    },
    "Convidado": {
      "type": "object",
      "properties": {
        "convidadoId": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "quantidadeConvidado": {
          "format": "int32",
          "type": "integer"
        },
        "casamentoId": {
          "format": "int32",
          "type": "integer"
        },
        "casamento": {
          "$ref": "#/definitions/Casamento"
        }
      }
    },
    "Decoracao": {
      "type": "object",
      "properties": {
        "decoracaoId": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "type": "string"
        },
        "item": {
          "type": "string"
        },
        "valor": {
          "format": "double",
          "type": "number"
        },
        "casamento": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Casamento"
          }
        }
      }
    },
    "Festa": {
      "type": "object",
      "properties": {
        "festaId": {
          "format": "int32",
          "type": "integer"
        },
        "nomeLocal": {
          "type": "string"
        },
        "capacidade": {
          "format": "int32",
          "type": "integer"
        },
        "horasContratadas": {
          "format": "date-span",
          "type": "string"
        },
        "valor": {
          "format": "double",
          "type": "number"
        },
        "buffetId": {
          "format": "int32",
          "type": "integer"
        },
        "bandaId": {
          "format": "int32",
          "type": "integer"
        },
        "buffet": {
          "$ref": "#/definitions/Buffet"
        },
        "banda": {
          "$ref": "#/definitions/Banda"
        },
        "casamento": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Casamento"
          }
        }
      }
    },
    "Operation": {
      "type": "object",
      "properties": {
        "value": { },
        "path": {
          "type": "string"
        },
        "op": {
          "type": "string"
        },
        "from": {
          "type": "string"
        }
      }
    },
    "Usuario": {
      "type": "object",
      "properties": {
        "loginId": {
          "format": "int32",
          "type": "integer"
        },
        "email": {
          "type": "string"
        },
        "nome": {
          "type": "string"
        },
        "senha": {
          "type": "string"
        },
        "temCasamento": {
          "format": "int32",
          "type": "integer"
        },
        "casamentoId": {
          "format": "int32",
          "type": "integer"
        },
        "casamento": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Casamento"
          }
        }
      }
    }
  }
}